+++ packages/api-explorer/src/ApiExplorer.tsx
@@ -25,8 +25,7 @@
  */
 
 import type { FC } from 'react'
-import React, { useReducer, useState, useEffect, useCallback } from 'react'
-import { useLocation } from 'react-router'
+import React, { useState, useEffect, useCallback } from 'react'
 import styled, { createGlobalStyle } from 'styled-components'
 import {
   Aside,
@@ -38,27 +37,25 @@ import {
   Page,
   Space,
 } from '@looker/components'
-import type { SpecList } from '@looker/sdk-codegen'
-import type { RunItSetter } from '@looker/run-it'
-import { funFetch, fallbackFetch, OAuthScene } from '@looker/run-it'
 import { FirstPage } from '@styled-icons/material/FirstPage'
 import { LastPage } from '@styled-icons/material/LastPage'
-
-import type { IEnvironmentAdaptor } from '@looker/extension-utils'
 import {
   registerEnvAdaptor,
   unregisterEnvAdaptor,
 } from '@looker/extension-utils'
-import { oAuthPath } from './utils'
+import { useSelector } from 'react-redux'
+import { useLocation } from 'react-router-dom'
+
+import type { IApixAdaptor } from './utils'
 import {
   Header,
   SideNav,
   ErrorBoundary,
-  Loader,
   SelectorContainer,
   HEADER_TOGGLE_LABEL,
+  Loader,
+  Banner,
 } from './components'
-import { specReducer, initDefaultSpecState, updateSpecApi } from './reducers'
 import { AppRouter } from './routes'
 import { apixFilesHost } from './utils/lodeUtils'
 import {
@@ -66,12 +63,15 @@ import {
   useSettingStoreState,
   useLodeActions,
   useLodesStoreState,
+  useSpecActions,
+  useSpecStoreState,
+  selectSpecs,
+  selectCurrentSpec,
 } from './state'
+import { getSpecKey, diffPath } from './utils'
 
 export interface ApiExplorerProps {
-  specs: SpecList
-  adaptor: IEnvironmentAdaptor
-  setVersionsUrl: RunItSetter
+  adaptor: IApixAdaptor
   examplesLodeUrl?: string
   declarationsLodeUrl?: string
   headless?: boolean
@@ -80,25 +80,21 @@ export interface ApiExplorerProps {
 const BodyOverride = createGlobalStyle` html { height: 100%; overflow: hidden; } `
 
 export const ApiExplorer: FC<ApiExplorerProps> = ({
-  specs,
   adaptor,
-  setVersionsUrl,
   examplesLodeUrl = 'https://raw.githubusercontent.com/looker-open-source/sdk-codegen/main/examplesIndex.json',
   declarationsLodeUrl = `${apixFilesHost}/declarationsIndex.json`,
   headless = false,
 }) => {
-  const { initialized } = useSettingStoreState()
+  useSettingStoreState()
   useLodesStoreState()
+  const { working, description } = useSpecStoreState()
+  const specs = useSelector(selectSpecs)
+  const spec = useSelector(selectCurrentSpec)
   const { initLodesAction } = useLodeActions()
   const { initSettingsAction } = useSettingActions()
-  const location = useLocation()
-  const oauthReturn = location.pathname === `/${oAuthPath}`
-  const [specState, specDispatch] = useReducer(
-    specReducer,
-    initDefaultSpecState(specs, location)
-  )
-  const { spec } = specState
+  const { initSpecsAction, setCurrentSpecAction } = useSpecActions()
 
+  const location = useLocation()
   const [hasNavigation, setHasNavigation] = useState(true)
   const toggleNavigation = (target?: boolean) =>
     setHasNavigation(target || !hasNavigation)
@@ -109,13 +105,23 @@ export const ApiExplorer: FC<ApiExplorerProps> = ({
     }
   }, [])
 
+  registerEnvAdaptor(adaptor)
+
   useEffect(() => {
-    registerEnvAdaptor(adaptor)
     initSettingsAction()
     initLodesAction({ examplesLodeUrl, declarationsLodeUrl })
 
+    const specKey = getSpecKey(location)
+    initSpecsAction({ specKey })
     return () => unregisterEnvAdaptor()
   }, [])
+
+  useEffect(() => {
+    const maybeSpec = location.pathname?.split('/')[1]
+    if (spec && maybeSpec && maybeSpec !== diffPath && maybeSpec !== spec.key) {
+      setCurrentSpecAction({ currentSpecKey: maybeSpec })
+    }
+  }, [location.pathname, spec])
 
   useEffect(() => {
     if (headless) {
@@ -128,27 +134,12 @@ export const ApiExplorer: FC<ApiExplorerProps> = ({
     }
   }, [headless, hasNavigationToggle])
 
-  useEffect(() => {
-    const loadSpec = async () => {
-      if (!spec.api) {
-        try {
-          const newSpec = { ...spec }
-          const api = await fallbackFetch(newSpec, funFetch)
-          if (api) {
-            spec.api = api
-            specDispatch(updateSpecApi(spec.key, api))
-          }
-        } catch (error) {
-          console.error(error)
-        }
-      }
-    }
-    if (!oauthReturn) {
-      loadSpec()
-    }
-  }, [spec, location])
-
   const themeOverrides = adaptor.themeOverrides()
+
+  let neededSpec = location.pathname?.split('/')[1]
+  if (!neededSpec || neededSpec === diffPath) {
+    neededSpec = spec?.key
+  }
 
   return (
     <>
@@ -156,18 +147,14 @@ export const ApiExplorer: FC<ApiExplorerProps> = ({
         loadGoogleFonts={themeOverrides.loadGoogleFonts}
         themeCustomizations={themeOverrides.themeCustomizations}
       >
-        {!initialized ? (
-          <Loader message="Initializing" themeOverrides={themeOverrides} />
+        {working || !neededSpec || neededSpec !== spec.key ? (
+          <Loader message={description} themeOverrides={themeOverrides} />
         ) : (
           <ErrorBoundary logError={adaptor.logError.bind(adaptor)}>
+            <Banner adaptor={adaptor} specs={specs} />
             <Page style={{ overflow: 'hidden' }}>
               {!headless && (
-                <Header
-                  specs={specs}
-                  spec={spec}
-                  specDispatch={specDispatch}
-                  toggleNavigation={toggleNavigation}
-                />
+                <Header spec={spec} toggleNavigation={toggleNavigation} />
               )}
               <Layout hasAside height="100%">
                 <AsideBorder
@@ -207,36 +194,22 @@ export const ApiExplorer: FC<ApiExplorerProps> = ({
                         <>
                           <Divider mb="u3" appearance="light" />
                           <SelectorContainer
+                            spec={spec}
                             ml="large"
                             mr="large"
-                            specs={specs}
-                            spec={spec}
-                            specDispatch={specDispatch}
                           />
                         </>
                       )}
                     </>
                   )}
-                  {hasNavigation && (
-                    <SideNav
-                      headless={headless}
-                      specs={specs}
-                      spec={spec}
-                      specDispatch={specDispatch}
-                    />
-                  )}
+                  {hasNavigation && <SideNav headless={headless} spec={spec} />}
                 </AsideBorder>
-                {oauthReturn && <OAuthScene />}
-                {!oauthReturn && spec.api && (
-                  <AppRouter
-                    api={spec.api}
-                    specKey={spec.key}
-                    specs={specs}
-                    toggleNavigation={toggleNavigation}
-                    adaptor={adaptor}
-                    setVersionsUrl={setVersionsUrl}
-                  />
-                )}
+                <AppRouter
+                  specKey={spec.key}
+                  api={spec.api!}
+                  specs={specs}
+                  toggleNavigation={toggleNavigation}
+                />
               </Layout>
             </Page>
           </ErrorBoundary>