+++ packages/api-explorer/src/state/specs/index.ts
@@ -23,4 +23,5 @@
  SOFTWARE.
 
  */
-export * from './spec'
+export * from './selectors'
+export * from './slice'