+++ packages/api-explorer/src/utils/path.ts
@@ -25,7 +25,8 @@
  */
 
 import type { ApiModel, IMethod, IType } from '@looker/sdk-codegen'
-import { firstMethodRef, Method } from '@looker/sdk-codegen'
+import { firstMethodRef } from '@looker/sdk-codegen'
+import type { Location as HLocation } from 'history'
 
 /**
  * Builds a path matching the route used by MethodScene
@@ -67,6 +68,12 @@ const getMethodTag = (api: ApiModel, methodName: string) => {
 }
 
 /**
+ * Is this item a method? Check without requiring `instanceof Method`
+ * @param item to check for method or type
+ */
+export const isMethod = (item: IMethod | IType) => 'params' in item
+
+/**
  * Return the tag for a give type
  * @param api Parsed api
  * @param type to tag
@@ -89,7 +96,7 @@ export const buildPath = (
   specKey: string
 ) => {
   let path
-  if (item instanceof Method) {
+  if (isMethod(item)) {
     const tag = getMethodTag(api, item.name)
     path = buildMethodPath(specKey, tag, item.name)
   } else {
@@ -97,4 +104,20 @@ export const buildPath = (
     path = buildTypePath(specKey, tag, item.name)
   }
   return path
+}
+
+/**
+ * Determine API specification keys from URL pattern
+ * @param location service to examine
+ */
+export const getSpecKey = (location: HLocation | Location): string | null => {
+  const pathname = location.pathname
+  let match
+  if (pathname.startsWith(`/${diffPath}`)) {
+    const pattern = new RegExp(`(?:/${diffPath})/(?<specKey>\\w+.\\w+)`)
+    match = pathname.match(pattern)
+  } else {
+    match = pathname.match(/\/(?<specKey>\w+\.\w+).*/)
+  }
+  return match?.groups?.specKey || null
 }